CORE-4200: Start refactoring flow API to align with review
