--- application/src/main/kotlin/net/corda/v5/application/flows/RpcStartFlowRequestParameters.kt
@@ -1,6 +1,0 @@
-package net.corda.v5.application.flows
-
-import net.corda.v5.base.annotations.CordaSerializable
-
-@CordaSerializable
-data class RpcStartFlowRequestParameters(val parametersInJson: String)
