+++ application/src/main/kotlin/net/corda/v5/application/flows/error/RetryFlowException.kt
@@ -1,4 +1,4 @@
-package net.corda.v5.application.flows
+package net.corda.v5.application.flows.error
 
 import net.corda.v5.base.exceptions.CordaRuntimeException
