+++ application/src/test/java/net/corda/v5/application/flows/FlowSessionJavaApiTest.java
@@ -1,6 +1,6 @@
 package net.corda.v5.application.flows;
 
-import net.corda.v5.application.identity.Party;
+import net.corda.v5.application.flows.messaging.FlowInfo;
 import net.corda.v5.base.types.MemberX500Name;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
