+++ application/src/test/java/net/corda/v5/application/flows/flowservices/FlowMessagingJavaApiTest.java
@@ -2,6 +2,7 @@ package net.corda.v5.application.flows.flowservices;
 
 import net.corda.v5.application.flows.FlowSession;
 import net.corda.v5.application.flows.UntrustworthyData;
+import net.corda.v5.application.flows.messaging.FlowMessaging;
 import net.corda.v5.base.types.MemberX500Name;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
